Revert "dash: add package, configure as CONFIG_SHELL"
This reverts commit 3b5986704e002602cf72c9e34206d2938dcc0a0b.

The config shell is set as an interpreter in libtool scripts
and as Linux has a hard limit of 127 bytes on the first line
in scripts (including the hashbang) this causes the build to
fail with a "bad interpreter" error if the path to the config
shell exceeds this limit.

These errors were first noticed on jenkins builds but can be
easily reproduced by building LibreELEC in a longer build dir.

eg RPi2 clean build failing on libtool invocation in kmod build:

./doltlibtool: /home/hias/rpi/libreelec-testing-with-a-long-directory-path-as-on-jenkins/build.LibreELEC-RPi2.arm-9.0-devel/kmod-24/.x86_64-linux-gnu/libtool: /home/hias/rpi/libreelec-testing-with-a-long-directory-path-as-on-jenkins/build.LibreELEC-RPi2.arm-9.0-devel/toolchain/bin/d: bad interpreter: No such file or directory

Signed-off-by: Matthias Reichl <hias@horus.com>