Enable OpenBOR in Emulationstation
OpenBOR only works with Pak files, if you have an extracted game you will need to create a pak first.

There is a bug/annoyance with the master.cfg file, if its the first time running a game it will start in a small window.

 You will need to go into options and change the video to full, after you do this, that game will always run in fullscreen.

 You need to do this the first time you run any pak.

 If master.cfg does not work, sound is weird or controller not working, you will need to use a keyboard to set your gamepad

 After you set up your gamne, copy the /storage/.config/openbor/Saves/{gamename}.cfg file to /storage/.config/openbor/master.cfg

master.cfg will only be copied the first time you run that particular game.